worked on room networking

+++ assignment-4/src/screens/game/dice.cpp
@@ -48,7 +48,7 @@ const std::vector<Die>& Dice::getDice() const { return dice; }
 
 void Dice::set(const std::vector<int>& faces) {
   auto it = dice.begin();
-  for (int i = 0; i < faces.size(); i++) {
+  for (size_t i = 0; i < faces.size(); i++) {
     dice[i].set(faces[i]);
     it++;
   }
